implement progress bar

+++ app/src/main/java/in/hahow/android_recruit_project/ui/home/component/CourseItem.kt
@@ -24,7 +24,7 @@ fun CourseItem(
     statusBackgroundColor: Color,
     progressString: String,
     progressBackgroundColor: Color,
-    progress: Float,
+    progress: Int,
     countDown: Long? = null,
 ) {
     Row(
